zswap: make shrinking memcg-aware
Implement a mechanism to allow a memcg specific reclaim. When a store
attempt hits an memcg limit, it now triggers a synchronous reclaim
attempt that, if successful, allows the new hotter page to be accepted
by zswap. If the store attempt instead hits the global zswap limit, it
will trigger an asynchronous reclaim attempt, in which an memcg is
selected for reclaim in a round-robin-like fashion.

Signed-off-by: Domenico Cerasuolo <cerasuolodomenico@gmail.com>
Co-developed-by: Nhat Pham <nphamcs@gmail.com>
Signed-off-by: Nhat Pham <nphamcs@gmail.com>