mm: zswap: fix potential memory corruption on duplicate store
While stress-testing zswap a memory corruption was happening when writing back
pages. __frontswap_store used to check for duplicate entries before attempting
to store a page in zswap, this was because if the store fails the old entry
isn't removed from the tree. This change removes duplicates entries in
zswap_store before the actual attempt.

Fixes: 42c06a0e8ebe ("mm: kill frontswap")
Signed-off-by: Domenico Cerasuolo <cerasuolodomenico@gmail.com>